Create pwm_control_DCMotor
This program is written to drive the two DC brushed motors on the TI Robot with PWM signals. The push button P1.1 will drive the motor forward while P1.4 will drive them backward. The duty cycle will be changed from 33% to 67% and 100% with each push on the switch, then go back to 33%. The motors only run when the switch is held.